// src/app/precisa/maestros/Parametros/model/parametros.ts
export class Parametros{

    constructor(){
        
    }
        CompaniaCodigo: string;
        AplicacionCodigo: string;
        DescripcionCorta: string;
        ParametroClave: string;
        DescripcionParametro: string;
        Explicacion: string;
        TipodeDatoFlag: string;
        Texto: string;
        Numero: number;
        Fecha: Date;
        FinanceComunFlag: string;
        Estado: string;
        UltimoUsuario: string;
        UltimaFechaModif: Date;
        ExplicacionAdicional: string;
        Texto1: string;
        Texto2: string;
        ESTADOdesc: string;
        UsuarioCreacion?:       string;
        FechaCreacion?:         Date;
        num: number;
}